+++ lib/src/rive_core/custom_property.dart
@@ -1,0 +1,8 @@
+import 'package:rive/src/generated/custom_property_base.dart';
+export 'package:rive/src/generated/custom_property_base.dart';
+
+/// A custom property value stored in the hierarchy with a boolean value.
+class CustomProperty extends CustomPropertyBase {
+  @override
+  void update(int dirt) {}
+}