+++ lib/src/rive_core/custom_property_boolean.dart
@@ -4,8 +4,5 @@ export 'package:rive/src/generated/custom_property_boolean_base.dart';
 /// A custom property value stored in the hierarchy with a boolean value.
 class CustomPropertyBoolean extends CustomPropertyBooleanBase {
   @override
-  void update(int dirt) {}
-
-  @override
   void propertyValueChanged(bool from, bool to) {}
 }